requirements.txt
# Python dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Azure SDK
azure-storage-blob==12.19.0
azure-identity==1.15.0
openai==1.12.0

# Monitoring
opencensus-ext-azure==1.1.13
opencensus-ext-flask==0.8.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
